refactor: move specific stuff out of the base method
