+++ lib/BaseWasmMainTemplatePlugin.js
@@ -78,38 +78,6 @@ function generateImportObject(module) {
 }
 
 class BaseWasmMainTemplatePlugin {
-	applyWeb(mainTemplate) {
-		const generateLoadBinaryCode = path =>
-			`fetch(${mainTemplate.requireFn}.p + ${path})`;
-
-		this.apply(mainTemplate, generateLoadBinaryCode);
-	}
-
-	applyNode(mainTemplate) {
-		const generateLoadBinaryCode = path => `new Promise(function (resolve, reject) {
-            var {readFile} = require("fs");
-            var {join} = require("path");
-
-            try {
-              readFile(join(__dirname, ${path}), function(err, buffer){
-                if (err) return reject(err);
-                
-                // Fake fetch response
-                resolve({
-                  arrayBuffer() {
-                    return Promise.resolve(buffer)
-                  }
-                });
-              });
-            } catch (err) {
-              reject(err);
-            }
-          });
-      `;
-
-		this.apply(mainTemplate, generateLoadBinaryCode);
-	}
-
 	apply(mainTemplate, generateLoadBinaryCode) {
 		mainTemplate.hooks.localVars.tap(
 			"BaseWasmMainTemplatePlugin",
