+++ lib/node/ReadFileCompileWasmTemplatePlugin.js
@@ -12,7 +12,33 @@ class ReadFileCompileWasmTemplatePlugin {
 		compiler.hooks.thisCompilation.tap(
 			"ReadFileCompileWasmTemplatePlugin",
 			compilation => {
-				new BaseWasmMainTemplatePlugin().applyNode(compilation.mainTemplate);
+				const generateLoadBinaryCode = path => `new Promise(function (resolve, reject) {
+                      var {readFile} = require("fs");
+                      var {join} = require("path");
+
+                      try {
+                        readFile(join(__dirname, ${
+													path
+												}), function(err, buffer){
+                          if (err) return reject(err);
+                          
+                          // Fake fetch response
+                          resolve({
+                            arrayBuffer() {
+                              return Promise.resolve(buffer)
+                            }
+                          });
+                        });
+                      } catch (err) {
+                        reject(err);
+                      }
+                    });
+                `;
+
+				new BaseWasmMainTemplatePlugin().apply(
+					compilation.mainTemplate,
+					generateLoadBinaryCode
+				);
 				new WasmModuleTemplatePlugin().apply(
 					compilation.moduleTemplates.javascript
 				);
