+++ lib/web/FetchCompileWasmMainTemplatePlugin.js
@@ -8,7 +8,13 @@ const BaseWasmMainTemplatePlugin = require("../BaseWasmMainTemplatePlugin");
 
 class FetchCompileWasmMainTemplatePlugin {
 	apply(mainTemplate) {
-		new BaseWasmMainTemplatePlugin().applyWeb(mainTemplate);
+		const generateLoadBinaryCode = path =>
+			`fetch(${mainTemplate.requireFn}.p + ${path})`;
+
+		new BaseWasmMainTemplatePlugin().apply(
+			mainTemplate,
+			generateLoadBinaryCode
+		);
 	}
 }
